backend/requirements.txt
fastapi
uvicorn
pydantic
googlemaps
reportlab
openpyxl
pandas
firebase-admin
python-dotenv
google-genai
